type of Reputation changed to Int

+++ src/main/java/com/javamentor/qa/platform/models/dto/user/UserDto.java
@@ -7,11 +7,10 @@ import lombok.Getter;
 @AllArgsConstructor
 public class UserDto {
     private Long id;
-
     private String email;
     private String fullName;
     private String imageLink;
     private String city;
-    private Long reputation;
+    private Integer reputation;
 
 }
